@@ -1,4 +1,14 @@
 export const environment = {
   production: false,
-  baseURL: 'http://localhost:8080/api/v1'
+  baseURL: 'http://localhost:8080/api/v1',
+  //baseURL: 'https://pocketpartners-backend.onrender.com/api/v1',
+  firebaseConfig: {
+    apiKey: "AIzaSyA4RzVQ5SXfExHEwo_0JAWZcn7q-rZxq7w",
+    authDomain: "agroconnect-demo-8c7ab.firebaseapp.com",
+    projectId: "agroconnect-demo-8c7ab",
+    storageBucket: "agroconnect-demo-8c7ab.appspot.com",
+    messagingSenderId: "617590434096",
+    appId: "1:617590434096:web:9957e257ce9221b7bc00f5",
+    measurementId: "G-004GHXNHG9"
+  }
 };
